Fix flakey tests in Safari
Tests were flakey in Safari because when you open a new tab it takes the camera away. So if you do this before you successfully publish then publishing times out because Mantis is waiting for video to start flowing. So to fix this I wait an extra couple of seconds after the publisher stops spinning. Longer term we should probably add an OT_publishing class to the publisher or something and we can check for that.