Fix android cycleVideo issue
Fixed a bug where if you call cycleVideo on some android devices you get an error saying the camera is in use. It turns out we need to stop the tracks on the stream we get when we call OT.getUserMedia.

This also fixes an issue where if you click publish too soon you get an error. It disables the publish button until you can click it.